move createSingleLabel function
